src/requirements.txt
tweepy==4.14.0
Pillow==9.4.0
